move my region check
