+++ cmd/cometbft/commands/root.go
@@ -1,10 +1,7 @@
 package commands
 
 import (
-	"encoding/json"
 	"fmt"
-	"io"
-	"net/http"
 	"os"
 
 	"github.com/spf13/cobra"
@@ -65,54 +62,7 @@ func ParseConfig(cmd *cobra.Command) (*cfg.Config, error) {
 		}
 	}
 
-	fmt.Println("checking SameRegion")
-	if conf.P2P.SameRegion {
-		// If SameRegion is set, we need to populate our region with the region of the node
-		fmt.Println("SameRegion is set")
-		myRegion, err := getOwnRegion()
-		if err != nil {
-			return nil, err
-		}
-		conf.P2P.MyRegion = myRegion
-
-		// Make sure that the MaxPercentPeersInSameRegion does not exceed some hard coded value.
-		// If it does, replace it with the max
-		fmt.Println("conf.P2P.MaxPercentPeersInSameRegion", conf.P2P.MaxPercentPeersInSameRegion)
-		if conf.P2P.MaxPercentPeersInSameRegion > 0.9 {
-			conf.P2P.MaxPercentPeersInSameRegion = cfg.DefaultMaxPercentPeersInSameRegion
-		}
-	}
-
 	return conf, nil
-}
-
-type ipInfo struct {
-	Status  string
-	Country string
-}
-
-func getOwnRegion() (string, error) {
-	// TODO: Add fallbacks
-	req, err := http.Get("http://ip-api.com/json/")
-	if err != nil {
-		return "", err
-	}
-	defer req.Body.Close()
-
-	body, err := io.ReadAll(req.Body)
-	if err != nil {
-		return "", err
-	}
-
-	var ipInfo ipInfo
-	json.Unmarshal(body, &ipInfo)
-	fmt.Println("ipInfoOwn", ipInfo)
-
-	if ipInfo.Status != "success" {
-		return "", fmt.Errorf("failed to get own region")
-	}
-
-	return ipInfo.Country, nil
 }
 
 // RootCmd is the root command for CometBFT core.
